Change notification policy for the sleeping blocks
Previously when nodes attained sleepStreak more than the
sleepStreakAlertLevel, monitor sent a notification. But if a node dies,
the monitor will send notifications forever because the node starts to stack
the sleepingStreak forever. Now the policy has been changed so that not
to send notifications forever but to send a notification when
sleepStreak starts to exceed the sleepStreakAlertLevel or has been
recovered from the problem.