[auto-self-nominate] Fix that transactions are pending in the mempool
Currently, if something went wrong with the previous tx and the node is not a
validator, the node keeps the failed transactions in the node's mempool. In
this case, the transactions sent through account_sendTransaction() are never
confirmed because of the seq.  This patch changes it to use
chain.sendTransaction() which calls chain.getSeq() RPC internally.